[#928] Split native_build.py's surviving logic into _native_build/
Move the Part 1 survivors (version-file generation, changelog/template
rendering, appdata.xml generation, Inno Setup script generation, and the
argv-dispatch orchestration) out of the single 599-line native_build.py
into a private `_native_build/` package (`version.py`, `templates.py`,
`meta.py`, `appdata.py`, `inno.py`, `cli.py`), with no behavior change.

`native_build.py` itself becomes a thin entry-point script that re-exports
`setup()` and `replace_placeholders()` (the latter still needed directly by
root `setup.py`), so `python native_build.py ...`/`./native_build.py ...`
and `import native_build` keep working unchanged.

`_native_build`/`_native_build.*` is added to pyproject.toml's package-find
exclude list so it's never bundled into the installed wheel, and
`DisplayCAL/setup.py`'s MANIFEST.in generator gets a
`recursive-include _native_build *.py` line so the sdist (which does need
it, to be buildable in isolation) still carries it. MANIFEST.in itself is
regenerated to pick up that line.
