+++ _native_build/__init__.py
@@ -1,0 +1,1 @@
+"""Packaging/freeze CLI implementation behind the `native_build.py` entry point."""
